requirements.txt
google-cloud-dataplex==2.2.2
google-api-python-client==2.139.0
google-auth==2.32.0
